* import.c (import_keys_internal): Missed one s/inp/inp2/.
* keylist.c (print_capabilities): Properly indicate per-key capabilities
of sign&encrypt primary keys that have secret-parts-missing (i.e. no
capabilities at all)

* mainproc.c (symkey_decrypt_sesskey): Fix compiler warning.

+++ g10/mainproc.c
@@ -245,7 +245,8 @@ symkey_decrypt_sesskey( DEK *dek, byte *sesskey, size_t slen )
     int n;
 
     if ( slen < 17 || slen > 33 ) {
-        log_error ( _("weird size for an encrypted session key (%d)\n"), slen);
+        log_error ( _("weird size for an encrypted session key (%d)\n"),
+		    (int)slen);
         return;   
     }
     hd = cipher_open( dek->algo, CIPHER_MODE_CFB, 1 );
